Update Tracking.cc
Appears that in localization mode if you cover the camera for a few seconds then uncover it, it does not properly relocalize. After some digging it seems like the mState variable is never updated from RECENTLY_LOST to LOST when in localization mode.

I took code from elsewhere in the program and added a check to update mState (may have done it wrong), but if mState does go to LOST while in localization mode it causes other issues, hangs on waiting for a map reset.

implemented jank solution

see https://github.com/UZ-SLAMLab/ORB_SLAM3/issues/377